AI-Image-Style-Transfer-Tool/CoAdapter/requirements.txt
transformers
diffusers
invisible_watermark
basicsr
einops
omegaconf
pytorch_lightning
kornia
gradio
opencv-python
pudb
imageio
imageio-ffmpeg
k-diffusion
webdataset
open-clip-torch
safetensors
timm
xformers
